Gjør om kafkaProdusent til å ta <Long, T> i stede for <String, T>

+++ mediator/src/main/kotlin/no.nav.dagpenger.arbeidssokerregister.mediator/connector/Model.kt
@@ -12,7 +12,7 @@ data class RecordKeyRequestBody(
 )
 
 data class RecordKeyResponse(
-    val key: Int,
+    val key: Long,
 )
 
 data class ArbeidssøkerperiodeResponse(
